# feat(aws): Add usage-file support for AWS FSx Windows File Server