--- a/infracost-usage-example.yml
+++ b/infracost-usage-example.yml
@@ -96,6 +96,9 @@ resource_usage:
   aws_instance.windows_instance:
     operating_system: windows # Override the operating system of the instance, can be: windows, suse, rhel.
 
+  aws_fsx_windows_file_system.my_system:
+    monthly_backup_storage_gb: 10000 # Monthly average amount of backup storage.
+
   aws_lambda_function.my_function:
     monthly_requests: 100000 # Monthly requests to the Lambda function.
     request_duration_ms: 500 # Average duration of each request in milliseconds.