--- a/infracost-usage-example.yml
+++ b/infracost-usage-example.yml
@@ -97,7 +97,7 @@ resource_usage:
     operating_system: windows # Override the operating system of the instance, can be: windows, suse, rhel.
 
   aws_fsx_windows_file_system.my_system:
-    monthly_backup_storage_gb: 10000 # Monthly average amount of backup storage.
+    backup_storage_gb: 10000 # Total storage used for backups in GB.
 
   aws_lambda_function.my_function:
     monthly_requests: 100000 # Monthly requests to the Lambda function.