Part of Issue #421 

With usage-file: 
```
  NAME                                          MONTHLY QTY  UNIT         PRICE   HOURLY COST  MONTHLY COST
  aws_fsx_windows_file_system.my_system                                                                      
  ├─ Throughput capacity                              1,024  MBps-months  4.5000       6.3123    4,608.0000  
  ├─ HDD storage                                        300  GB-months    0.0250       0.0103        7.5000  
  └─ Backup storage                                  10,000  GB-months    0.0500       0.6849      500.0000  
  Total                                                                                7.0075    5,115.5000
```

Without usage-file:
```
  NAME                                          MONTHLY QTY  UNIT         PRICE   HOURLY COST  MONTHLY COST
  aws_fsx_windows_file_system.my_system                                                                      
  ├─ Throughput capacity                              1,024  MBps-months  4.5000       6.3123    4,608.0000  
  ├─ HDD storage                                        300  GB-months    0.0250       0.0103        7.5000  
  └─ Backup storage                                       -  GB-months    0.0500            -             -  
  Total                                                                                6.3226    4,615.5000  
```